Refactor RatingCalculator to utilize dedicated match and track rating calculators
- Introduced `MatchRatingCalculator` and `TrackRatingCalculator` to encapsulate match and track rating logic, improving code organization and maintainability.
- Updated the `reset` method to clear state through the new calculators.
- Modified `getRankings` to gather unique users from the calculators, enhancing flexibility in user management.
- Streamlined the `processTimeEntry` and `processMatch` methods to delegate processing to the respective calculators, ensuring separation of concerns.
